+++ CHANGES.md
@@ -2,4 +2,6 @@
 # 0.1.0 / 2020-03-25
 
   * Initial Release
-  * code migrated from pypi packages luriegold, kfactor, onepara 
+  * code migrated from pypi packages luriegold, illmat
+  * randcorr function added
+